[mkcal] a new API to read/write into a single DB file. Contributes to JB#60800
Dissociate database read/write from the calendar
storage. In practice, it is introducing a new class
that do DB read/write with API centred on Incidence::List
and Notebook::List.

This new class is still using the SqliteFormat underneath
but is fully compatible with incidences sharing the same
UID between notebooks.

SqliteStorage has been rewritten to use this new class
but it still cannot handle incidences with the same UID
because it is still storing the data into ExtendedCalendar.