[mkcal] Add listing methods for sync in the backend.
Move the incidence listing methods from
SqliteStorage to the backend. These methods
allow to have a more optimized way to
list incidences in the sync plugins. But
as mentioned in the documentation, they
may be also a bit fragile since they rely
on user-modifiable fields.

A more robust way to implement the delta
calculation in sync (but less optimized)
is to rely on incidences(nbuid) and
deletedIncidences(nbuid) which list all
incidences to be compared to the server
listing.